worker/requirements.txt
firebase-admin>=6.0.0
pyyaml>=6.0
